requirements.txt
flask
Pillow
numpy
